requirements-dev.txt
fab
Fabric3
factory-boy==2.11.1
mkdocs==1.0.4
pytest==3.8.2
pytest-django==3.4.3
